Generate XML and HTML coverage reports when running android tests
The Jacoco task from the Android Gradle plugin is completely custom. It
generates automatically reports but it does not support cross-modules
coverage.
As the android tests are in a dedicated module, by default, the Android
Jacoco will not gather the coverage of the lib module. The workaround is
to make the sources of the lib belong to the sources of the test module.
This mess up the IDE, so this workaround is only applied when Gradle is
invoked with a dedicated parameter.